requirements.txt
# MCP SDK
mcp>=1.0.0

# Webex Bot SDK
webexteamssdk>=1.6.1

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# HTTP Client
requests>=2.31.0
httpx>=0.25.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Async support
asyncio-throttle>=1.0.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0

# Logging
structlog>=23.2.0
